Added tests, refactored to jsx
