+++ client/src/components/RatingsAndReviews/Ratings/Ratings.jsx
