+++ package.json
@@ -4,7 +4,7 @@
   "description": "FEC Project",
   "main": "server.js",
   "scripts": {
-    "test": "jest",
+    "test": "jest --env=jsdom",
     "server": "nodemon server.js",
     "transpile": "npx webpack --config webpack.config.js --mode development -w"
   },
@@ -24,12 +24,13 @@
     "@babel/plugin-transform-runtime": "^7.15.0",
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-react": "^7.14.5",
+    "@testing-library/jest-dom": "^5.14.1",
     "axios": "^0.21.4",
     "babel-jest": "^27.1.1",
     "babel-loader": "^8.2.2",
     "express": "^4.17.1",
-    "jest": "^27.1.1",
     "moment": "^2.29.1",
+    "node-fetch": "^3.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-test-renderer": "^17.0.2",
@@ -37,9 +38,12 @@
     "webpack": "^5.52.1"
   },
   "devDependencies": {
+    "@testing-library/react": "^12.1.0",
     "css-loader": "^6.2.0",
+    "jest": "^27.1.1",
     "less": "^4.1.1",
     "less-loader": "^10.0.1",
+    "regenerator-runtime": "^0.13.9",
     "style-loader": "^3.2.1",
     "webpack-cli": "^4.8.0"
   }
