+++ server.js
@@ -1,9 +1,8 @@
 const path = require("path")
-const express = require("express"); // npm installed
+const express = require("express");
 
 const app = express();
 
 app.use(express.static(path.join(__dirname, "/client/dist")));
-// other configuration...
 
 app.listen(3000);
