+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path')
 
 module.exports = {
-  entry: path.resolve(__dirname, 'client/src', 'index.js'),
+  entry: path.resolve(__dirname, 'client/src', 'index.jsx'),
   output: {
     path: path.resolve(__dirname, 'client/dist'),
     filename: 'bundle.js'
@@ -29,5 +29,9 @@ module.exports = {
         }
       }
     ]
+  },
+  resolve: {
+    modules: ["node_modules",path.resolve(__dirname, "app")],
+    extensions: [".jsx", ".js", ".json", ".css"]
   }
 }
